+++ bigtree/node/node.py
@@ -213,6 +213,19 @@ class Node(BaseNode):
 
         print_tree(self, **kwargs)
 
+    def __getitem__(self, child_name: str) -> T:
+        """Get child by name identifier
+
+        Args:
+            child_name (str): name of child node
+
+        Returns:
+            (Self): child node
+        """
+        from bigtree.tree.search import find_child_by_name
+
+        return find_child_by_name(self, child_name)  # type: ignore
+
     def __repr__(self) -> str:
         """Print format of Node